+++ joulescope_ui/software_update.py
@@ -1,4 +1,4 @@
-# Copyright 2019-2023 Jetperch LLC
+# Copyright 2019-2026 Jetperch LLC
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -15,6 +15,7 @@
 """Check for software updates"""
 
 from PySide6 import QtWidgets, QtCore
+import datetime
 import requests
 import json
 import threading
@@ -120,6 +121,8 @@ def _platform_name():
     if psys == 'Windows':
         if platform.machine() == 'AMD64':
             return 'win10_x86_64'  # for both win10 and win11
+        elif platform.machine() == 'ARM64':
+            return 'win11_arm64'  # Windows on ARM
     elif psys == 'Linux':
         # assume all Linux is the supported Ubuntu version for now
         return 'ubuntu_24_04_x86_64'
@@ -344,13 +347,20 @@ class SoftwareUpdateDialog(QtWidgets.QDialog):
 
     @QtCore.Slot(int)
     def _on_finish(self, value):
+        record_dismissal = False
         if not is_release:
             _log.info('software update: not a release')
+            record_dismissal = True
         elif value == QtWidgets.QDialog.DialogCode.Accepted:
             _log.info('software update: update now')
             self._pubsub.publish('registry/ui/actions/!close', {'software_update': self._info})
         else:
             _log.info('software update: later')
+            record_dismissal = True
+        if record_dismissal:
+            self._pubsub.publish(
+                'registry/app/settings/software_update_check_previous',
+                datetime.datetime.utcnow().isoformat())
         if self._done_action is not None:
             self._pubsub.publish(*self._done_action, defer=True)
         self.close()